Issue #548
Here is a patch that allow to keep control on the maximum size of the logfiles.
Two new attributes have been added to the "node" element of roslaunch XML files :
* max_logfile_size : a positive or null integer. This is the max size in byte that the logfile of a node will count.
* logfile_count : a positive or null integer. This is how much logfile will be rotated. To understand how this work, read the RotatingFileHandler of logging.handlers python module.

Unit Tests :
* I did some change in test-node-valid.xml, to ensure that adding the two new attributes don't break anything.
* I added two test XML files which are tested in a new unit test test_xmlloader.py.

That seem quite complexe to automate the testing of the feature I introduce in this patch (launching a node, and test that a rotating log file is created). So I only test it "by hand" analyzing a log folder when launching some big projects I do in my company, and some trivial launch files...

indigo-rethink cherry-pick Conflicts:
	tools/roslaunch/src/roslaunch/nodeprocess.py