Re #577: Signal slot & onRetryTimer callback fix
This issue manifested with random coredumps in the onRetryTimer()
callback in transport_publisher_link.cpp. These crashes were infrequent,
and extremely difficult to reproduce. They appeared to happen when
many Subscriptions were being destroyed, which made it likely to be
a destruction scheduling anomaly.

It turned there were two issues at play: the onRetryTimer function did
not having a valid reference to the TransportPublusherLink class, and
the onConnectionDropped function was not registered to the
boost::signals2::signal in such a way that tracked the existence of
the underlying ROS Subscriber connection. The first issue (which was
uncovered in #577) was resolved by calling getInternalTimerManger()::add
with shared_from_this instead of an empty shared pointer. Testing this
fix then revealed the second issue. Initially, the onConnectionDropped
seemed to be fixed by keeping track of the TransportPublisherLink
class with a shared_from_this pointer, though it did not pass all of
the ros_comm tests. However, changing anything this deep in ROS runs the
risk of unintended consequences, and upon further testing, we discovered
that ros_controls/realtime_tools::RealTimePublishers were not being
properly notified when connections were dropped, and entering deadlock
conditions.

Taking inspiration from the boost::signals2 Automatic Connection Management
tutorial [1], we then used signals2::slot_type and track() mechanisms in order
to maintain a reference to the TransportPublisherLink class during execution
of the onConnectionDropped() code. This was functionally similar to the
previous approach, with the added benefit that boost::signals2 handles
all shared_from_this() reference counting under the hood, without preventing
destruction of the parent class (which we believe to be the source of the
unintended RealTimePublisher deadlocking). Upon the tracked object's
destruction, the signal2 library automatically disconnects the signal.

Reproduction of this issue was incredibly difficult, since it required
the timing of the class destruction to perfectly coincide with the
execution of the callback function using that class. We found this could
only be reliably reproduced when running overnight simulations of hundreds
of Subscribers by Rethink Robotics Automated QA team. The occurrence of
this bug was producing roughly 1 to 5 coredumps per night in this testing
environment. Since this patch was applied, zero coredumps and
RealTimePublisher deadlocks have been seen over the course of a several
days of testing.

Major credit to Jeff Adams (@jeffadamsc) for pursuing the shared_from_this
callback connection strategy, and for QA allowing us to run these tests on
their nightly runs.

[1] http://www.boost.org/doc/libs/1_54_0/doc/html/signals2/tutorial.html#signals2.tutorial.connection-management